+++ examples/custom_render/src/main.rs
@@ -13,7 +13,7 @@ fn main() {
 
 /// Draw solid background to window buffer.
 fn draw(mut wrapper_query: Query<&mut PixelsWrapper>) {
-    let Ok(mut wrapper) = wrapper_query.get_single_mut() else {
+    let Ok(mut wrapper) = wrapper_query.single_mut() else {
         return;
     };
     let frame = wrapper.pixels.frame_mut();
@@ -23,7 +23,7 @@ fn draw(mut wrapper_query: Query<&mut PixelsWrapper>) {
 
 /// Custom render system.
 pub fn render(wrapper_query: Query<&PixelsWrapper>) {
-    let Ok(wrapper) = wrapper_query.get_single() else {
+    let Ok(wrapper) = wrapper_query.single() else {
         return;
     };